Advanced RAG with LangChain, Groq, and FAISS/requirements.txt

python-dotenv
streamlit
langchain_community
langchain
faiss-cpu
langchain-groq
langchain_google_genai